0.16.0

+++ pkg/model/ch_config_const.go
@@ -63,21 +63,27 @@ const (
 )
 
 const (
-	// Default ClickHouse docker image to be used
+	// defaultClickHouseDockerImage specifies default ClickHouse docker image to be used
 	defaultClickHouseDockerImage = "yandex/clickhouse-server:latest"
 
-	// Default BusyBox docker image to be used
+	// defaultBusyBoxDockerImage specifies default BusyBox docker image to be used
 	defaultBusyBoxDockerImage = "busybox"
 
-	// Name of container within Pod with ClickHouse instance. Pod may have other containers included, such as monitoring
-	ClickHouseContainerName    = "clickhouse"
+	// Name of container within Pod with ClickHouse instance.
+	// Pod may have other containers included, such as monitoring, logging
+
+	// ClickHouseContainerName specifies name of the clickhouse container in the pod
+	ClickHouseContainerName = "clickhouse"
+	// ClickHouseLogContainerName specifies name of the logger container in the pod
 	ClickHouseLogContainerName = "clickhouse-log"
 )
 
 const (
-	chPortNumberMustBeAssignedLater = 0
+	// chPortNumberMustBeAssignedLater value means that port
+	// is not assigned yet and is expected to be assigned later.
+	chPortNumberMustBeAssignedLater = int32(0)
 
-	// ClickHouse open ports
+	// ClickHouse open ports names and values
 	chDefaultTCPPortName               = "tcp"
 	chDefaultTCPPortNumber             = int32(9000)
 	chDefaultHTTPPortName              = "http"
@@ -87,6 +93,7 @@ const (
 )
 
 const (
+	// zkDefaultPort specifies Zookeeper default port
 	zkDefaultPort = 2181
 	// zkDefaultRootTemplate specifies default ZK root - /clickhouse/{namespace}/{chi name}
 	zkDefaultRootTemplate = "/clickhouse/%s/%s"
